+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added: Debug log display option to toggle between time sequence and grouped by request views.
 - Added: Debug log autorefresh option.
 - Added: Workflow engine logs "Engine finished processing" on shutdown when debug is enabled.
+- Added: Added cache for workflow execution context variables to increase performance when resolving variables repeated times (Issue #1581).
 - Changed: Debug log now defaults to grouped by request display.
 - Changed: Move the scheduled actions "Run" button to it's own column, (Issue #1496).
 - Changed: Workflow debug log messages rewritten for clarity and consistency (format, parentheses for details, [Workflow] prefix).